+++ code/src/Machine.h
@@ -5,6 +5,7 @@
 #ifndef MRP_MACHINE_H
 #define MRP_MACHINE_H
 
+#include <stdint.h>
 #include "sets.h"
 
 class Machine {
@@ -26,18 +27,18 @@ public:
 
     void add_process(unsigned int process_id);
     void remove_process(unsigned int process_id);
-    void add_capacity(unsigned int capacity);
-    void add_safety_capacity(unsigned int safety_capacity);
+    void add_capacity(uint32_t capacity);
+    void add_safety_capacity(uint32_t safety_capacity);
     void add_mmc(unsigned int mmc);
     unsigned int get_id();
     unsigned int get_location_id();
     unsigned int get_neighborhood_id();
-    void set_usage(unsigned int resource_id, unsigned long int usage);
-    void set_transient_usage(unsigned int resource_id, unsigned int usage);
-    unsigned long int get_transient_usage(unsigned int resource_id);
-    unsigned long int get_usage(unsigned int resource_id);
-    unsigned long int get_capacity(unsigned int resource_id);
-    unsigned long int get_scapacity(unsigned int resource_id);
+    void set_usage(unsigned int resource_id, uint32_t usage);
+    void set_transient_usage(unsigned int resource_id, uint32_t usage);
+    uint32_t get_transient_usage(unsigned int resource_id);
+    uint32_t get_usage(unsigned int resource_id);
+    uint32_t get_capacity(unsigned int resource_id);
+    uint32_t get_scapacity(unsigned int resource_id);
     unsigned int get_machine_move_cost(unsigned int machine_id);
     void print();
     bool has_process(unsigned int process_id);
